fix: Circular dependency
