+++ src/redteam/types.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 
 import type { ApiProvider, ProviderOptions } from '../types/providers';
 import type { Plugin, Severity } from './constants';
-import { isValidPolicyId } from './plugins/policy/utils';
+import { isValidPolicyId } from './plugins/policy/validators';
 
 // Modifiers are used to modify the behavior of the plugin.
 // They let the user specify additional instructions for the plugin,
